Don't show button outline when button isn't pressed
Increase the contrast of the icons/text on the button by having it be white
on black background instead of white on gray background.

This will be helpful when we introduce pixel shifting in the next few commits,
as now the off-centering of icons is going to be harder to notice.