Introduce automatic pixel shifting
The panel in the touchbar is an OLED screen and pixels on it will suffer
degradation over time. Degradation effects are going to be worst on bright,
white pixels that remain at the same position forever. This applies to the
icons of our buttons.

To lower the possibilty of a burn-in here as much as we can, introduce some
subtle pixel shifting.

Here's how it works:

- Use the existing TIMEOUT_MS poll() timeout and (temporarily) allow smaller
  timeouts for smooth sub-pixel transitions to make the pixel shift completely
  invisible

- We have lots of space in the horizontal direction, so there we shift pixels
  by the whole amount that's necessary to ensure no pixel remains "always on"

- The background boxes around each button are also shifted in horizontal
  direction as otherwise the icons appear off-center horizontally, for this
  the boxes need to be reduced in width a bit so they don't overflow the panel

- In the vertical direction, icons appear off-center on the whole touchbar
  really quickly, so only do very limited movement here, that's not very
  effective but better than nothing

- Hitboxes for touch input remain the same and are not affected by pixel
  shifting

This starts a pixel shift by one full pixel in x and y direction every 10
seconds, transitioning slowly between the states on a sub-pixel level for
4 seconds. Once the pixel shift reaches the final left or right edge of
the shifting area, we'll leave it at that for 50 seconds to reduce stress
on pixels in the center of the shifting area a bit more.